longhui

+++ src/main/java/com/qg/service/NoteCommentService.java
@@ -23,7 +23,7 @@ public class NoteCommentService {
 	 * @return true false
 	 */
 	public boolean addNoteComment(NoteCommentModel noteComment){
-		return noteCommentDao.addNoteComment(noteComment);
+		return (new NoteService().existNote(noteComment.getNoteId()))?noteCommentDao.addNoteComment(noteComment):false;
 	}
 	/***
 	 * 根据id获取留言评论
@@ -41,8 +41,8 @@ public class NoteCommentService {
 	 */
 	public boolean deleteComment(int commentId,int userId) {
 		//判断权限后删除(评论者和留言板本人才可以删除)
-		if (this.getNoteCommentById(commentId).getCommenterId()==userId||
-				userId==new NoteService().getNoteById(this.getNoteCommentById(commentId).getNoteId()).getTargetId())
+		if (this.getNoteCommentById(commentId)!=null&&this.getNoteCommentById(commentId).getCommenterId()==userId||
+				this.getNoteCommentById(commentId)!=null&&userId==new NoteService().getNoteById(this.getNoteCommentById(commentId).getNoteId()).getTargetId())
 			return noteCommentDao.deleteComment(commentId);
 		else
 			return false;
